metrics: Move SparkSampler to separate module
