+++ gprofiler/spark/mode.py
@@ -1,0 +1,7 @@
+#
+# Copyright (c) Granulate. All rights reserved.
+# Licensed under the AGPL3 License. See LICENSE.md in the project root for license information.
+#
+SPARK_DRIVER_MODE = "driver"
+SPARK_YARN_MODE = "yarn"
+SPARK_MESOS_MODE = "mesos"
